searchbar works in console.log and layout for the output for search are ready
