+++ src/components/Recipes.js
@@ -4,7 +4,6 @@ const Recipes = (props) => {
   return (
     <>
       <div className="Recipes d-flex flex-wrap justify-content-center">
-      
         {props.recipes.map((element, index) => {
           return (
             <RecipesCards
@@ -15,7 +14,6 @@ const Recipes = (props) => {
           );
         })}
       </div>
-      
     </>
   );
 };
